+++ python/test/test_Parser.py
@@ -4,7 +4,7 @@
 #-------------------------------------------------------------------------------
 import unittest
 
-from tweedledum.BoolFunctionCompiler import BitVec, Parser, BoolFunction
+from tweedledum.bool_function_compiler import BitVec, Parser, BoolFunction
 
 def id(a: BitVec(1)) -> BitVec(1):
     return a
